ajout getScore de joueur, suppression de fonctions/attributs inutiles, correction de certains bugs

+++ Code/ia.cpp
@@ -19,7 +19,7 @@ void IA::ajouterTuile(Tuile* t) {
 bool IA::tuileContientPlace(Tuile* t) {
     if (!t) return false;
     for (size_t i = 0; i < t->getNbHexagones(); ++i) {
-        Hexagone* h = t->getHexagone(i);
+        Hexagone* h = t->getHexagone((int)i);
         if (h->getType() == Place) {
             return true;
         }
@@ -71,7 +71,7 @@ int IA::calculerScoreIA() const {
 
     for (Tuile* t : tuilesAcquises) {
         for (size_t i = 0; i < t->getNbHexagones(); i++) {
-            Hexagone* h = t->getHexagone(i);
+            Hexagone* h = t->getHexagone((int)i);
             Type type = h->getType();
 
             if (type == Carriere) {
